fix: add reset-offline-peers command
